added final touchup and edited summary msg
