+++ db.py
@@ -7,5 +7,8 @@ db = client.get_database("hackandroll")
 
 records = db.student_records
 
+# Restart DB in a clean state
+# records. delete_many({})
+
 
 print(records.count_documents({}))
